+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.3.0] - TBD
 ### Added
 - Added code snippets for all tunable nodes in XML (T, E, U, L, V, I, M, C).
+- Added `projectMetaData` and the ability to use variables in S4TK config.
 ### Fixed
 - Fixed issue where formatting a file would not respect the "Tab Size" setting.
 ### Removed